+++ packages/babel-plugin-transform-classes/src/vanilla.js
